requirements.txt
torch==2.0.0
torchvision==0.15.0
bitarray
git+https://github.com/facebookresearch/detectron2.git
fvcore
ipdb
matplotlib
numpy
opencv-python
packaging
git+https://github.com/cocodataset/panopticapi.git
Pillow
pycocotools
scipy
setuptools
terminaltables
timm
tqdm
transformers
ftfy 
regex 
tqdm
cython
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/segment-anything.git

